refactoring/2152 Use namespace-based APIs instead of Tenant-based APIs
Intermediate commit - quite a few tests missing for appropriate coverage.

* Rewritten failing AccountControllerTest tests
* Renamed TenantServiceIntegrationTest to Namespace... and commented out all tests / re-written some tests (needs more)
* Simplified adding user to namespace (user was expressed both as path variable and payload - leaving only payload for consistency)
* Fixed bug in DefaultUserAccountService when adding non-existing user to namespace (chances of occurrences higher in the future when APIs are V1) --> was throwing NPE instead of ultimately responding back with failure. Still ugly now, but does not crash at least.
