+++ repository/repository-server/src/test/java/org/eclipse/vorto/repository/server/it/AllTests.java
@@ -7,7 +7,7 @@ import org.junit.runners.Suite.SuiteClasses;
 @RunWith(Suite.class)
 @SuiteClasses({AttachmentsControllerIntegrationTest.class, ModelControllerIntegrationTest.class,
     ModelSearchControllerIntegrationTest.class, RepositoryJavaClientTest.class,
-    TenantServiceIntegrationTest.class})
+    NamespaceServiceIntegrationTest.class})
 public class AllTests {
 
 }
