+++ repository/repository-server/src/test/java/org/eclipse/vorto/repository/web/AccountControllerTest.java
@@ -72,7 +72,7 @@ public class AccountControllerTest extends AbstractIntegrationTest {
         .andExpect(status().isOk())
         .andReturn();
     // quick, but dirty
-    String expectedUsernameKV = "\"username\":\"testUser\"";
+    String expectedUsernameKV = "\"userId\":\"testUser\"";
     assertTrue(result.getResponse().getContentAsString().contains(expectedUsernameKV));
   }
 
@@ -85,7 +85,8 @@ public class AccountControllerTest extends AbstractIntegrationTest {
         .andExpect(status().isOk())
         .andReturn();
     // quick, but dirty
-    String expectedUsernameKV = "\"username\":\"testUser\"";
+    String expectedUsernameKV = "\"userId\":\"testUser\"";
+    String debug = result.getResponse().getContentAsString();
     assertTrue(result.getResponse().getContentAsString().contains(expectedUsernameKV));
   }
 
@@ -98,7 +99,7 @@ public class AccountControllerTest extends AbstractIntegrationTest {
         .andExpect(status().isOk())
         .andReturn();
     // quick, but dirty
-    String expectedUsernameKV = "\"username\":\"testUser\"";
+    String expectedUsernameKV = "\"userId\":\"testUser\"";
     assertEquals(result.getResponse().getContentAsString(), "[]");
   }
